MOD: modify type and oal function

+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/ServiceInstanceJVMThread.java
@@ -54,11 +54,11 @@ public class ServiceInstanceJVMThread extends Source {
     private String serviceId;
     @Getter
     @Setter
-    private int liveCount;
+    private long liveCount;
     @Getter
     @Setter
-    private int daemonCount;
+    private long daemonCount;
     @Getter
     @Setter
-    private int peakCount;
+    private long peakCount;
 }
